Update logic for synchronizing system information
The current approach of synchronizing system information
by deleting the existing data in the database and replacing
it with new data had issues with difficulty in tracking changes
and a lack of adherence to RESTful principles.
The proposed solution involved refactoring commit.py to enable
synchronization on a per-system information basis.