Fix the location of global lock
Place the global lock previously used in
command.py within commit.py.
It has been decided that placing the global
lock in commit.py is more appropriate, so it
has been fixed.
The usage of additional threads for sync
in command.py has been replaced with simple
function calls.
Creating additional threads for sync might be
redundant, so it has been fixed.
Change in URL due to update in alpacon-server.