Refactor makeArchive()
Previously, use cd command to navigate to
the working directory, but when executing in
sh -c environment, replace spaces in the path
with "// " for compatibility. This reduced readability.
To address this, implemente a solution using cmd.Dir
and cmd.Args to change the working directory,
allowing us to input paths with spaces directly.