Add InitFtpLogger()
To ensure consistent logging format
with the existing logs, InitFtpLogger()
was added to a new process executing
Web FTP commands.
As the new process has reduced permissions on
/var/alpamon/alpamon.log, InitLogger alone was
insufficient to unify the log structure.
To further consolidate logs from the new process
into the existing log file, Unix domain sockets
should be implemented in the future.