Add collectorArgs and collectConf type
Add collectConf type to facilitate the
retrieval of metric collection configuration
from alpacon-server.
Add collectorArgs type to encapsulate the
parameters required for collector instantiation.
The initialization process for the collector
has been streamlined into the
InitCollector() function.
Inside InitCollector(), fetch the metric
collection configuration and construct
the object to be passed as an argument.