Update delUser command
Add PurgeHomeDirectory to deleteUserData
to support system user purge.
Update delUser so that when a system user
is deleted without purge, their home directory
is backed up to /home/deleted_users.
The backup is in the format of {username\}_{timestamp}.
Update delUser to transfer ownership of these backed-up
directories to the root user for security reasons .